Use the actual bone model
