+++ RedefineGuide/Extensions/SceneKitExtensions.swift
@@ -1,0 +1,47 @@
+//
+//  SceneKitExtensions.swift
+//  RedefineGuide
+//
+//  Created by Stephen Potter on 4/10/24.
+//
+
+import Foundation
+import SceneKit
+
+extension SCNNode {
+    
+    /// Sets the euler angles based on degree inputs.
+    /// - Parameters:
+    ///   - node: The SCNNode to modify.
+    ///   - x: The rotation around the x-axis in degrees.
+    ///   - y: The rotation around the y-axis in degrees.
+    ///   - z: The rotation around the z-axis in degrees.
+    func rotate(x: CGFloat = 0, y: CGFloat = 0, z: CGFloat = 0) {
+        let radiansPerDegree = CGFloat.pi / 180
+        self.eulerAngles = SCNVector3(x * radiansPerDegree, y * radiansPerDegree, z * radiansPerDegree)
+    }
+    
+    /// Scales the node so it has the specified width in centimeters.
+    func scaleToWidth(_ centimeters: Float) {
+        let (min, max) = self.boundingBox
+        let currentWidth = max.x - min.x
+        let targetWidth = Float(centimeters) / 100.0
+        let scaleFactor = targetWidth / currentWidth
+        self.scale = SCNVector3(scaleFactor, scaleFactor, scaleFactor)
+    }
+}
+
+extension SCNVector3 {
+    func normalized() -> SCNVector3 {
+        let length = sqrt(x * x + y * y + z * z)
+        return SCNVector3(x / length, y / length, z / length)
+    }
+    
+    static func +(lhs: SCNVector3, rhs: SCNVector3) -> SCNVector3 {
+        return SCNVector3(lhs.x + rhs.x, lhs.y + rhs.y, lhs.z + rhs.z)
+    }
+    
+    static func *(vector: SCNVector3, scalar: Float) -> SCNVector3 {
+        return SCNVector3(vector.x * scalar, vector.y * scalar, vector.z * scalar)
+    }
+}
